Fix Sets "tuple" example because it's misleading
This example is incorrect. The statement

```
Set(1,2,3) - (1,2)
```

for example does not use a tuple at all, it's the same thing as this:

```
Set(1,2,3).-(1,2)
```

Which is calling the overloaded version of the `-` method of `Set` which
accepts two initial arguments plus a varadiac list of however many
elements you want which allows you to do things like

```
Set(1,2,3,4) - (1,2,3,4,5) // etc
```

No where in this are tuples ever actually used. And to say so (as it
does currently) is misleading to programmers trying to use
scala-exercises as a way of learning the standard library. If you DID
try to use a tuple it would not work:

```
Set(1,2,3) - ((1,2)) //this will not compile
```