Respect backwards compat in SetsSpec for tuple correction
@raulraja pointed out that it'd be a good idea to keep the number of
arguments in the exercise I'm clarifying the same so that users who have
already answered the exercise still see it marked as correct.

So this commit removes the additional test meant to help hammer the
point about varadiac arguments and sets the arguments to the test to be
the same as before, and moves the note about varadiac-ness to the
comment on the existing test itself.